ISSUE-8 created components for loginButton, loginScreen, password and user
