+++ .expo/packager-info.json
@@ -1,7 +1,7 @@
 {
-  "expoServerPort": 19000,
-  "packagerPort": 19001,
-  "packagerPid": 27051,
+  "expoServerPort": 19001,
+  "packagerPort": 19002,
+  "packagerPid": 27778,
   "expoServerNgrokUrl": null,
   "packagerNgrokUrl": null,
   "ngrokPid": null
