+++ src/components/loginButton/Styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+module.exports = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    button: {
+      backgroundColor: "#3c6fc1",
+      paddingVertical: 20,
+      alignItems: "center",
+      justifyContent: "center",
+      marginTop: 30,
+    },
+    buttonText: {
+      color: "#FFF",
+      fontSize: 18,
+    }
+})
