+++ src/components/loginScreen/Styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+module.exports = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    markWrap: {
+        flex: 1,
+        paddingVertical: 30,
+    },
+    mark: {
+        width: null,
+        height: null,
+        flex: 1,
+    },
+    wrapper: {
+        paddingVertical: 30,
+    }
+})
