+++ src/components/newAccount/Styles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+module.exports = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    signupWrap: {
+        backgroundColor: "transparent",
+        flexDirection: "row",
+        alignItems: "center",
+        justifyContent: "center",
+        color: 'black'
+    },
+    accountText: {
+        color: "#000"
+    },
+    signupLinkText: {
+        color: "#000",
+        marginLeft: 5,
+    }
+})
