+++ src/components/password/Password.js
@@ -1,43 +1,23 @@
 import React, { Component } from 'react';
-import { View, StyleSheet, TextInput } from 'react-native';
+import { View, TextInput } from 'react-native';
+import Styles from './Styles.js';
 import { Ionicons } from '@expo/vector-icons';
 
 class Password extends Component {
     render() {
         return (
-            <View style={styles.inputPasswordWrap}>
-                <View style={styles.iconWrap}>
+            <View style={Styles.inputPasswordWrap}>
+                <View style={Styles.iconWrap}>
                     <Ionicons name='ios-lock' size={20} />
                 </View>
                 <TextInput
                     placeholderTextColor="#000"
                     placeholder="Senha"
-                    style={styles.input}
+                    style={Styles.input}
                     secureTextEntry/>
             </View>
         );
     }
 }
-
-const styles = StyleSheet.create(
-    {
-    inputPasswordWrap: {
-        flexDirection: "row",
-        marginVertical: 10,
-        height: 40,
-        borderBottomWidth: 1,
-        borderBottomColor: "#CCC"
-      },
-      input: {
-        flex: 1,
-        paddingHorizontal: 10,
-      },
-      iconWrap: {
-        paddingHorizontal: 7,
-        alignItems: "center",
-        justifyContent: "center",
-      },
-    }
-)
 
 export default Password;
