+++ src/components/user/User.js
@@ -1,42 +1,22 @@
 import React, { Component } from 'react';
 import { View, StyleSheet, TextInput } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
+import Styles from './Styles.js';
 
 class User extends Component {
     render() {
         return (
-            <View style={styles.inputUserWrap}>
-                <View style={styles.iconWrap}>
+            <View style={Styles.inputUserWrap}>
+                <View style={Styles.iconWrap}>
                     <Ionicons name='ios-person' size={20} />
                 </View>
                 <TextInput
                     placeholder="Usuário"
                     placeholderTextColor="#000"
-                    style={styles.input}/>
+                    style={Styles.input}/>
             </View>
         );
     }
 }
-
-const styles = StyleSheet.create(
-    {
-    inputUserWrap: {
-        flexDirection: "row",
-        marginVertical: 10,
-        height: 40,
-        borderBottomWidth: 2,
-        borderBottomColor: "#CCC"
-      },
-      input: {
-        flex: 1,
-        paddingHorizontal: 10,
-      },
-      iconWrap: {
-        paddingHorizontal: 7,
-        alignItems: "center",
-        justifyContent: "center",
-      },
-    }
-)
 
 export default User;
